extract repair_whitelist
