+++ validator/src/commands/mod.rs
@@ -4,6 +4,7 @@ pub mod exit;
 pub mod monitor;
 pub mod plugin;
 pub mod repair_shred_from_peer;
+pub mod repair_whitelist;
 pub mod set_identity;
 pub mod set_log_filter;
 pub mod staked_nodes_overrides;
